Move register and cancel to helper functions
